Add X86Lab::Snapshot type to represent snapshot of a Vm's state.
Each snapshot contains a copy of the VM's state (registers and memory)
at some point in the execution.
For now, each snapshot contain a full copy of the VM's physical memory
at that point in time, hence no deduplication is performed. However this
can easily be added later since each snapshot get a pointer to the
previous snapshot, essentially building a linked-list.