Re-work testing framework.
Now each test must be declared with the DECLARE_TEST macro which
automatically registers the test to be run when calling runAllTests().
Tests have been changed to avoid using the old assert() macro and
instead use TEST_ASSERT() which throws exceptions when asserts are
failing. This allows to catch such exception and pretty print failing
tests.