Re-work how physical memory is allocated for the VM.
Vm::addPhysicalMemory no longer take an offset argument, the allocated
physical memory is appended to the Vm's physical memory.
Additionally, it makes sure that it is not using more than the maximum
number of slots available for the Vm and has the option to render the
allocated physical memory read-only for the guest.
Page tables structures allocated for long-mode paging are now mapped to
read-only physical memory.