Refactor page table initialization in 64-bit.
Incidentally this also fixed an issue that was not detected by any of
the tests: Attempting to write anything in memory in LongMode would lead
to an error because page tables were allocated in read-only memory and
therefore the vcpu failed to set the dirty bit.

This issue did not happen with the accessed bit, although the vcpu
obviously access pages when executing code. This could be due to
information caching, as described in Intel's manual vol3 "4.8 ACCESSED
AND DIRTY FLAGS".