Kvm: Fix issue with call to KVM_GET_SUPPORTED_CPUID failing due to destination struct too small.
KVM_GET_SUPPORTED_CPUID needs the destination struct to be large enough
to contain the result (e.g. enough entry in variable sized array / ptr).
Until now the number of entry was hard-coded to 64, which is not enough
on some machine.
Change the wrapper around KVM_GET_SUPPORTED_CPUID to use a suitable
sized struct computed by increasing the size of the struct until the
ioctl does not return error E2BIG.