VM: Fix test failure of testSetRegisters due to VM entry error.
This failure was due to the fact that the GDT and IDT bases were loaded
with random, non-canonical addresses. On Intel hosts, this triggers an
error upon VM entry, presumably due to the Guest State Area not passing
the state (although the documentation does not mention anything
regarding checking the bases). This issue was not triggered on AMD
hosts.
Use canonical addresses for GDT and IDT bases in testSetRegisters to
avoid this issue.