Fix issue occuring when un-focusing the window where the process would use 100% cpu.
This was due to an underflow in a loop condition when drawing the memory
window leading to the process spending all it's time making draw calls.

Also fix an issue with the stack window where it would be shrunk
horizontally upon loosing focus. Due to missing call to
SetNextWindowSize with auto-fit values.