GUI+Backend: Add a way to change the starting cpu mode of the VM.
The GUI now displays three radio buttons, one for each cpu mode (16, 32
or 64 bit mode) that choosed the cpu mode in which the VM starts.
Changing the starting mode resets the VM into that mode. The "Reset"
button resets the VM in the currently selected mode.

This commit also removes the command line options --16, --32 and --64
which were used until now to choose the starting mode. These options
might be re-added in the future.